Task folders opened
Task 2 improvements;

- UI created,
- Task requirements completed,
- First manual test completed.

Now;

The user can set an alarm and see a result notification about alarm setting state result. The app will send a notification via the Local API when the timer is complete. In this way, the application always sends a notification, even if the application were be killed.

Requirements upgrades;
- Audio requirement (increasing) will be added,
- Interface bug fixes will be solved,
- Code review will be completed.